Move tracing module into cmd, since it should not be part of the library code.
