+++ sdk/nodejs/runtime/index.ts
@@ -29,4 +29,3 @@ export * from "./resource";
 export * from "./rpc";
 export * from "./settings";
 export * from "./stack";
-export * from "./tracing";
